4-backend/analytics-py-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
motor==3.3.2
celery==5.3.4
aioredis==2.0.1
python-dotenv==1.0.0